kernel: move error funcs to error.c
